requirements.txt
alembic==1.8.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
black==22.6.0
boto3==1.24.56
botocore==1.27.56
click==8.1.3
colorama==0.4.5
config==0.5.1
factory-boy==3.2.1
Faker==14.1.0
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
functions==0.7.0
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
password-strength==0.0.3.post2
pathspec==0.9.0
phonenumbers==8.12.53
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
s3transfer==0.6.0
schemas==0.7.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
utils==1.0.1
uuid==1.30
Werkzeug==2.2.2
zipp==3.8.1
